+++ ci/test_mysql.py
@@ -16,18 +16,18 @@ DATABASES = {
     "default": {
         "ENGINE": "django.db.backends.mysql",
         "NAME": "django_default",
-        "USER": "django",
         "HOST": "127.0.0.1",
-        "PASSWORD": "secret",
-        "TEST": {"CHARSET": "utf8mb4", "COLLATION": "utf8mb4_general_ci"},
+        "USER": "scott",
+        "PASSWORD": "tiger",
+        "TEST": {"CHARSET": "utf8mb3", "COLLATION": "utf8mb3_general_ci"},
     },
     "other": {
         "ENGINE": "django.db.backends.mysql",
         "NAME": "django_other",
-        "USER": "django",
         "HOST": "127.0.0.1",
-        "PASSWORD": "secret",
-        "TEST": {"CHARSET": "utf8mb4", "COLLATION": "utf8mb4_general_ci"},
+        "USER": "scott",
+        "PASSWORD": "tiger",
+        "TEST": {"CHARSET": "utf8mb3", "COLLATION": "utf8mb3_general_ci"},
     },
 }
 
@@ -37,3 +37,5 @@ SECRET_KEY = "django_tests_secret_key"
 PASSWORD_HASHERS = [
     "django.contrib.auth.hashers.MD5PasswordHasher",
 ]
+
+DEFAULT_AUTO_FIELD = "django.db.models.AutoField"