+++ doc/conf.py
@@ -22,6 +22,13 @@ import os
 
 # -- General configuration -----------------------------------------------------
 
+nitpick_ignore = [
+    ("py:class", "datetime.date"),
+    ("py:class", "datetime.time"),
+    ("py:class", "datetime.datetime"),
+]
+
+
 # If your documentation needs a minimal Sphinx version, state it here.
 # needs_sphinx = "1.0"
 
@@ -42,8 +49,8 @@ source_suffix = ".rst"
 master_doc = "index"
 
 # General information about the project.
-project = "MySQLdb"
-copyright = "2012, Andy Dustman"
+project = "mysqlclient"
+copyright = "2023, Inada Naoki"
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
@@ -93,7 +100,7 @@ pygments_style = "sphinx"
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
-html_theme = "default"
+html_theme = "sphinx_rtd_theme"
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the