+++ src/MySQLdb/release.py
@@ -1,0 +1,3 @@
+__author__ = "Inada Naoki <songofacandy@gmail.com>"
+__version__ = "2.2.4"
+version_info = (2, 2, 4, "final", 0)